fix: remove requestforms references (MTI-5812) (#24)
This covers both the main site as well as the LP.Most of this one was fairly straightforward in removing and replacing references. However, one of the pages stands out just a bit in comparison to the other. On `wp-content/themes/uwa/teacher-connect.php` there were two references to request forms. One for mobile and one for desktop. Upon inspecting these in production, they are already using TLH Forms rather than requestforms, and have been named `tlh-form-mobile` and `tlh-form`. I've reproduced this in the code in order to follow suit with what is working currently.